refactor interactor structure
