+++ Cargo.toml
@@ -9,16 +9,14 @@ members = [
   "fee-market",
   "fee-market/meta",
   "header-verifier",
-  "header-verifier/meta", 
+  "header-verifier/meta",
   "enshrine-esdt-safe",
-  "enshrine-esdt-safe/interactor",
   "enshrine-esdt-safe/meta",
   "token-handler",
   "token-handler/meta",
   "testing-sc",
   "testing-sc/meta",
   "sovereign-forge",
-  "sovereign-forge/interactor",
   "sovereign-forge/meta",
   "header-verifier/meta",
   "sov-esdt-safe",
@@ -28,5 +26,5 @@ members = [
   "testing-sc",
   "testing-sc/meta",
   "interactor",
-  "common/token-whitelist"
+  "common/token-whitelist",
 ]
