+++ common/common-interactor/src/constants.rs
