+++ common/common-interactor/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod common_interactor_sovereign;
 pub mod constants;
 pub mod interactor_config;
 pub mod interactor_state;
