--- common/interactor/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "interactor"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-toml = "0.8.6"
-
-[dependencies.multiversx-sc]
-version = "=0.57.1"
-
-[dependencies.multiversx-sc-snippets]
-version = "=0.57.1"
-
-[dependencies.serde]
-version = "1.0"
-features = ["derive"]
