--- enshrine-esdt-safe/interactor/Cargo.toml
@@ -1,63 +1,0 @@
-[[bin]]
-name = "enshrine_esdt_safe_interactor"
-path = "src/enshrine_esdt_safe_interactor_main.rs"
-
-[package]
-name = "enshrine_esdt_safe_interactor"
-version = "0.0.0"
-authors = ["you"]
-edition = "2021"
-publish = false
-
-[lib]
-path = "src/enshrine_esdt_safe_interactor.rs"
-
-[dependencies]
-toml = "0.8.6"
-
-[dependencies.enshrine-esdt-safe]
-path = ".."
-
-[dependencies.structs]
-path = "../../common/structs"
-
-[dependencies.setup-phase]
-path = "../../common/setup-phase"
-
-[dependencies.token-whitelist]
-path = "../../common/token-whitelist"
-
-[dependencies.interactor]
-path = "../../common/interactor"
-
-[dependencies.utils]
-path = "../../common/utils"
-
-[dependencies.fee-market]
-path = "../../fee-market"
-
-[dependencies.header-verifier]
-path = "../../header-verifier"
-
-[dependencies.token-handler]
-path = "../../token-handler"
-
-[dependencies.proxies]
-path = "../../common/proxies"
-
-[dependencies.multiversx-sc-snippets]
-version = "=0.57.1"
-
-[dependencies.multiversx-sc]
-version = "=0.57.1"
-
-[dependencies.clap]
-version = "4.4.7"
-features = ["derive"]
-
-[dependencies.serde]
-version = "1.0"
-features = ["derive"]
-
-[features]
-chain-simulator-tests = []
