+++ interactor/src/enshrine_esdt_safe/mod.rs
@@ -1,0 +1,1 @@
+pub mod enshrine_esdt_safe_interactor;
