+++ interactor/src/sovereign_forge/mod.rs
@@ -1,0 +1,1 @@
+pub mod sovereign_forge_interactor_main;
