--- sovereign-forge/interactor/Cargo.toml
@@ -1,42 +1,0 @@
-[[bin]]
-name = "forge-rust-interact"
-path = "src/interactor_main.rs"
-
-[package]
-name = "forge-rust-interact"
-version = "0.0.0"
-authors = ["you"]
-edition = "2021"
-publish = false
-
-[lib]
-path = "src/interact.rs"
-
-[dependencies]
-toml = "0.8.6"
-
-[dependencies.sovereign-forge]
-path = ".."
-
-[dependencies.multiversx-sc-snippets]
-version = "0.57.1"
-
-[dependencies.multiversx-sc]
-version = "0.57.1"
-
-[dependencies.structs]
-path = "../../common/structs"
-
-[dependencies.proxies]
-path = "../../common/proxies"
-
-[dependencies.clap]
-version = "4.4.7"
-features = ["derive"]
-
-[dependencies.serde]
-version = "1.0"
-features = ["derive"]
-
-[features]
-chain-simulator-tests = []
