--- sovereign-forge/interactor/config.toml
@@ -1,7 +1,0 @@
-
-chain_type = 'simulator'
-gateway_uri = 'http://localhost:8085'
-#
-# chain_type = 'real'
-# gateway_uri = 'https://devnet-gateway.multiversx.com'
-
