--- sovereign-forge/interactor/src/interactor_main.rs
@@ -1,7 +1,0 @@
-use forge_rust_interact::sovereign_forge_cli;
-use multiversx_sc_snippets::imports::*;
-
-#[tokio::main]
-async fn main() {
-    sovereign_forge_cli().await;
-}
